refactor: move data into global data folder
