+++ src/pages/projects/index.njk
@@ -13,7 +13,7 @@ layout: layout.njk
   logistics applications…
 </p>
 
-{% for projectType in breakdown %}
+{% for projectType in global.breakdown %}
   <div>
     <h2 class="cmp-project__group-heading">{{ projectType[0] }}</h2>
     <ul class="cmp-project__list">
